+++ notifier/sns/notifier_test.go
@@ -236,7 +236,8 @@ func TestNotifier(t *testing.T) {
 				}
 			},
 			notified: true,
-			log: `level=INFO msg="Subscribed sqs queue to sns topic." queue=https://sqs.us-west-2.amazonaws.com/123456789012/MyQueue topic=topic
+			log: `level=INFO msg="Start watching SNS topic." topic=topic queue=https://sqs.us-west-2.amazonaws.com/123456789012/MyQueue
+level=INFO msg="Received messages from SNS topic." topic=topic count=1
 level=WARN msg="No loader to process message." msg=message
 `,
 		},
@@ -304,7 +305,8 @@ level=WARN msg="No loader to process message." msg=message
 				}
 			},
 			notified: true,
-			log: `level=INFO msg="Subscribed sqs queue to sns topic." queue=https://sqs.us-west-2.amazonaws.com/123456789012/MyQueue topic=topic
+			log: `level=INFO msg="Start watching SNS topic." topic=topic queue=https://sqs.us-west-2.amazonaws.com/123456789012/MyQueue
+level=INFO msg="Received messages from SNS topic." topic=topic count=1
 level=WARN msg="Fail to process message." msg=message loader=loader error="process message error"
 `,
 		},
@@ -442,7 +444,8 @@ level=WARN msg="Fail to process message." msg=message loader=loader error="proce
 				}
 			},
 			notified: true,
-			log: `level=INFO msg="Subscribed sqs queue to sns topic." queue=https://sqs.us-west-2.amazonaws.com/123456789012/MyQueue topic=topic
+			log: `level=INFO msg="Start watching SNS topic." topic=topic queue=https://sqs.us-west-2.amazonaws.com/123456789012/MyQueue
+level=INFO msg="Received messages from SNS topic." topic=topic count=1
 level=WARN msg="Fail to delete sqs queue." queue=https://sqs.us-west-2.amazonaws.com/123456789012/MyQueue error="operation error SQS: DeleteQueue, delete queue error"
 `,
 		},
@@ -548,7 +551,8 @@ level=WARN msg="Fail to delete sqs queue." queue=https://sqs.us-west-2.amazonaws
 				}
 			},
 			notified: true,
-			log: `level=INFO msg="Subscribed sqs queue to sns topic." queue=https://sqs.us-west-2.amazonaws.com/123456789012/MyQueue topic=topic
+			log: `level=INFO msg="Start watching SNS topic." topic=topic queue=https://sqs.us-west-2.amazonaws.com/123456789012/MyQueue
+level=INFO msg="Received messages from SNS topic." topic=topic count=1
 level=WARN msg="Fail to unsubscribe sns topic." topic=topic error="operation error SNS: Unsubscribe, unsubscribe error"
 `,
 		},
@@ -601,7 +605,7 @@ level=WARN msg="Fail to unsubscribe sns topic." topic=topic error="operation err
 				}
 			},
 			notified: false,
-			log: `level=INFO msg="Subscribed sqs queue to sns topic." queue=https://sqs.us-west-2.amazonaws.com/123456789012/MyQueue topic=topic
+			log: `level=INFO msg="Start watching SNS topic." topic=topic queue=https://sqs.us-west-2.amazonaws.com/123456789012/MyQueue
 level=WARN msg="Fail to receive sqs message." queue=https://sqs.us-west-2.amazonaws.com/123456789012/MyQueue error="operation error SQS: ReceiveMessage, receive message error"
 `,
 		},
@@ -666,7 +670,8 @@ level=WARN msg="Fail to receive sqs message." queue=https://sqs.us-west-2.amazon
 				}
 			},
 			notified: true,
-			log: `level=INFO msg="Subscribed sqs queue to sns topic." queue=https://sqs.us-west-2.amazonaws.com/123456789012/MyQueue topic=topic
+			log: `level=INFO msg="Start watching SNS topic." topic=topic queue=https://sqs.us-west-2.amazonaws.com/123456789012/MyQueue
+level=INFO msg="Received messages from SNS topic." topic=topic count=1
 level=WARN msg="Fail to delete sqs message." queue=https://sqs.us-west-2.amazonaws.com/123456789012/MyQueue error="operation error SQS: DeleteMessageBatch, delete message error"
 `,
 		},
@@ -694,10 +699,10 @@ level=WARN msg="Fail to delete sqs message." queue=https://sqs.us-west-2.amazona
 			)
 			assert.NoError(t, err)
 
-			buf := &buffer{}
 			opts := []ksns.Option{
 				ksns.WithAWSConfig(cfg),
 			}
+			buf := &buffer{}
 			if testcase.log != "" {
 				opts = append(opts, ksns.WithLogHandler(logHandler(buf)))
 			}