+++ CHANGELOG.md
@@ -10,6 +10,7 @@ to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 ### Added
 
 - Support change notification via SNS topic (#267).
+- Support change notification via PubSub topic (#294).
 
 ## [1.0.0] - 2024-03-16